okay this is kinda a big one. So the parsing is more or less complete. Will need to have someone show me some edge cases to work with but it should check maps well. I did a bunch of refactoring and moving around functions to make it more clear. IE. adding all the find functions into one find_functions.c file
