+++ maps/OK/map.cub
@@ -1,10 +1,11 @@
-NO                      textures/wolfenstein/grey_stone.png
+NO textures/texture1.png
 
-SO textures/wolfenstein/blue_stone.png
+SO textures/texture2.png
 
-WE textures/wolfenstein/red_brick.png
+WE textures/texture3.png
 
-EA textures/wolfenstein/wood.xpm
+EA textures/texture4.png
+
 
 F 18,53,25
 
@@ -12,9 +13,9 @@ C 153,204,255
 
       1111111
 111111111111                  111111111
-100000000011111111111111111111100000001
-10000000000000000000000000000000000S001
-111111111101111111111111111111100000001
+100000000011111111111111111111111100001
+100000000000000000000000000000000000N01
+111111111111111111111111111111100000001
 100000000001                  111111111
 111111111101
 110000000001        111111
