--- maps/OK/map1.cub
@@ -1,22 +1,0 @@
-NO textures/texture1.png
-
-SO textures/texture2.png
-
-WE textures/texture3.png
-
-EA textures/texture4.png
-
-
-F 18,53,25
-
-C 153,204,255
-
-      1111111
-111111111111                  111111111
-100000000011111111111111111111111100001
-100000000000000000000000000000000000W01
-111111111101111111111111111111100000001
-100000000001                  111111111
-111111111101
-110000000001        111111
-111111111111
