+++ srcs/main.c
@@ -6,7 +6,7 @@
 /*   By: jbremser <jbremser@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/25 15:51:35 by jbremser          #+#    #+#             */
-/*   Updated: 2024/11/14 17:51:44 by jbremser         ###   ########.fr       */
+/*   Updated: 2024/12/18 15:56:23 by jbremser         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,13 +23,12 @@ int main(int argc, char **argv)
 		return (handle_error(EXIT_ARG_COUNT_ERROR, game));
 	if (!handle_error(parse_args(argv, game), game))
 		return (1);
-	printf("\nhere\n");
-	if (!handle_error(init_game(game), game))
-		return (1);
-	start_game(game);
-	// if (game)
-	// 	free_game_struct(game);
+	printf("\nPARSING COMPLETE!\n");
+	// if (!handle_error(init_game(game), game))
+	// 	return (1);
+	// start_game(game);
+	if (game)
+		free_game_struct(game);
 	return (0);
 
 }
-
